+++ GoToDnSpy/ReleaseNotes.md
@@ -1,5 +1,10 @@
 # Release notes
 
+- 1.0.3
+    + Added navigate support for Roslyn many others `*DeclarationSyntax`  
+        (less "is not a valid identifier" errors)
+    + Add support for referenced project output (i.e. your referenced class library project)
+      (no more 'not found' errors on this type of reference)
 - 1.0.2
     + Added navigate support for Roslyn `PropertyDeclarationSyntax`  
         (less "is not a valid identifier" errors)